haitham 28-3

+++ app/Http/Controllers/Admin/AdminController.php
@@ -126,8 +126,12 @@ class AdminController extends Controller
     public function destroy($id)
     {
         $admin = Admin::find($id);
+        if( $id != 9 ){
         $admin->delete();
-        return redirect()->route('admins.index')->with('message','تم جذف اليوزر بنجاح');
+            return redirect()->route('admins.index')->with('message','تم جذف اليوزر بنجاح');
+        }elseif( $id == 9 ){
+            return redirect()->route('admins.index')->with('warning','لا يمكن حذف هذا اليوزر !');
+        }
 
     }
 }
